Added SoapCore tuner service
